+++ airflow/providers/mysql/provider.yaml
@@ -47,7 +47,6 @@ versions:
 dependencies:
   - apache-airflow>=2.3.0
   - apache-airflow-providers-common-sql>=1.3.1
-  - mysql-connector-python>=8.0.11
   - mysqlclient>=1.3.6
 
 integrations:
@@ -87,3 +86,8 @@ transfers:
 connection-types:
   - hook-class-name: airflow.providers.mysql.hooks.mysql.MySqlHook
     connection-type: mysql
+
+additional-extras:
+  - name: mysql-connector-python
+    dependencies:
+      - mysql-connector-python>=8.0.11
