+++ scripts/in_container/verify_providers.py
@@ -150,12 +150,12 @@ KNOWN_COMMON_DEPRECATED_MESSAGES: set[str] = {
     "Implementing implicit namespace packages (as specified in PEP 420) is "
     "preferred to `pkg_resources.declare_namespace`",
     "This module is deprecated. Please use `airflow.providers.cncf.kubernetes.operators.pod` instead.",
-    "urllib3 (1.26.6) or chardet (5.1.0)/charset_normalizer (2.0.12) doesn't match a supported version!",
-    "urllib3 (1.26.9) or chardet (5.1.0)/charset_normalizer (2.0.12) doesn't match a supported version!",
     "This operator is deprecated. Please use `GoogleDisplayVideo360CreateQueryOperator`",
     "This operator is deprecated. Please use `GoogleDisplayVideo360RunQueryOperator`",
     "This operator is deprecated. Please use `GoogleDisplayVideo360RunQuerySensor`",
     "This operator is deprecated. Please use `GoogleDisplayVideo360DownloadReportV2Operator`",
+    "urllib3 (1.26.6) or chardet (5.1.0)/charset_normalizer (2.0.12) doesn't match a supported version!",
+    "urllib3 (1.26.9) or chardet (5.1.0)/charset_normalizer (2.0.12) doesn't match a supported version!",
 }
 
 # The set of warning messages generated by direct importing of some deprecated modules. We should only
